refactor: theme injection
